refactor: improve class names
