+++ index.js
@@ -1,11 +1,11 @@
 import os from 'node:os'
 import timer from 'timers/promises'
-import task from './src/task/index.js'
 import prompt from './src/prompt/index.js'
-import Uptimer from './src/uptimer/index.js'
 import threadpool from './src/threadpool/index.js'
-import Collector from './src/collector/index.js'
+import runTask from './src/run-task/index.js'
+import Uptimer from './src/uptimer/index.js'
 import Scheduler from './src/scheduler/index.js'
+import { MetricsCollector } from './src/metrics/index.js'
 
 const dyno = async (taskFn, { 
   parameters, 
@@ -28,9 +28,9 @@ const dyno = async (taskFn, {
     })
 
     const abortctrl = new AbortController()
-    const collector = new Collector()
-    const uptimer = new Uptimer()
+    const collector = new MetricsCollector()
     const scheduler = new Scheduler(parameters)
+    const uptimer = new Uptimer()
     const threads = await threadpool.fork(
       // @TODO document the following line intention
       typeof taskFn === 'function' ? process.argv[1] : taskFn, { 
@@ -40,8 +40,8 @@ const dyno = async (taskFn, {
     
     process.start()
     scheduler.start(threads)
-    uptimer.start()
     collector.start(threads, onTick.bind(this))
+    uptimer.start()
 
     try {
       await Promise.race([
@@ -62,7 +62,7 @@ const dyno = async (taskFn, {
     return collector.result()
   }
   
-  return task(taskFn)
+  return runTask(taskFn)
 }
 
-export { dyno, task }
+export { dyno, runTask as task }
