+++ src/metrics/classes/rec-histogram.js
